refactor(agent): remove `semantic_guard.py` and merge its functionality into `agent.py`, updating related unit tests and value objects
